Parplexity correctly formatted export
