+++ src/manifest.json
@@ -1,8 +1,8 @@
 {
   "manifest_version": 3,
-  "name": "Save my Phind - AI Conversation Exporter",
-  "version": "1.9.9",
-  "description": "🚀 Export your Phind.com search threads into markdown files!\n",
+  "name": "Save My AI - Phind and Perplexity Exporter",
+  "version": "2.0.1",
+  "description": "🚀 Export your Phind.com and Perplexity.ai search threads into markdown files!\n",
   "icons": {
     "16": "assets/icons/icon-16.png",
     "48": "assets/icons/icon-48.png",
